requirements.txt
boto3>=1.34
openai>=1.30
pydantic>=2.0
PyYAML>=6.0
requests>=2.31
tenacity>=8.2
